Changed data format of sonar output to 8 bit mono, vertically flipped output of FLS, fixed depth camera near/far planes, fixed usage of depth clamping.

+++ Library/include/sensors/vision/FLS.h
@@ -136,7 +136,7 @@ namespace sf
         void InitGraphics();
         
         OpenGLFLS* glFLS;
-        GLfloat* sonarData;
+        GLubyte* sonarData;
         GLubyte* displayData;
         glm::vec2 range;
         Scalar gain;
